Fixed game handling on memory error allocating second player
